Map: use sprites for level selection buttons
Drawing the buttons directly to the raster is much more expensive, using
a significant amount of CPU power per tick.

An attempt at using sprites had already been made, but with only two
16x16 images without numbers. This time, each button has its own image
loaded into VRAM.

When in the 'map scene, all 17 images for inactive buttons are kept in
VRAM, while only one image for the active button is loaded.