Editor, Level: move all editing code to its own module
Rather than using two fake entities for the 'edit cursor' and 'edit
sidebar', it is better to have a separate module that handles all of the
editor system.

Most of the code in 'editor.c' is copied from 'edit-cursor.c' and
'edit-sidebar.c' and then adapted to work outside of the entity system.