Player: fix obstacles not breaking when they should
After a recent change to the initial movement speed, wood obstacles
could no longer be broken from a distance of 4 tiles.

Since depending on the movement speed to determine if an obstacle should
break is a fragile logic that requires precise fine-tuning, it is best
to instead use the distance traveled, measured in tiles.