Level, Player: do not call 'level_load' while entities are ticking
The player falls from four tiles above its spawn location. Since
'level_load' was being called within 'tick_entities', the solid
partition of the player was updated *after* reloading the level,
causing a ghost reference to the player to be stuck in that tile.

It just happens that Level 9 has a platform exactly four tiles above the
player's spawn position.